* misc.h, misc.c, sievecmd.c: 	Removed read_script_file and extended read_from_stream to handle 	reading until an EOF is found. Improved handling of EOF marker.
git-svn-id: https://svn.ic-s.nl/svn/dbmail/trunk/dbmail@2286 7b491191-dbf0-0310-aff6-d879d4d69008

+++ misc.h
@@ -96,7 +96,7 @@ int ci_write(FILE * fd, char * msg, ...);
 int num_from_imapdate(const char *date);
 
 
-int read_from_stream(FILE * instream, char **m_buf, size_t maxlen);
+int read_from_stream(FILE * instream, char **m_buf, int maxlen);
 int find_bounded(const char * const value, char left, char right,
 		char **retchar, size_t * retsize, size_t * retlast);
 int zap_between(const char * const instring, signed char left, signed char right,
